refactor: Refactor startup process to delay setting window size.
Setting initial window size moved into draw_frame, and called
after fonts are loaded. So initial window size matches requested
geometry.

This also fixes missing startup screen, as UiCommand::Resize is not
called on startup.

+++ src/bridge/ui_commands.rs
@@ -16,8 +16,8 @@ use crate::windows_utils::{
 pub enum UiCommand {
     Quit,
     Resize {
-        width: u32,
-        height: u32,
+        width: u64,
+        height: u64,
     },
     Keyboard(String),
     MouseButton {
